+++ src/xliff-parser/xliff-parser.ts
@@ -1,43 +1,29 @@
 import { ValidationError, XMLParser, XMLValidator } from 'fast-xml-parser';
-import { IXliffTag, IXliff, XliffTagName } from '../models/xliff';
-
-interface IXMLParserTextNode {
-  '#text': string;
-}
-
-interface IXMLParserAttributes {
-  ':@'?: {
-    [attributeName: string]: string;
-  };
-}
-
-type IXMLParserChild = IXMLParserAttributes & {
-  [tagName: string]:
-    | (IXMLParserChild | IXMLParserTextNode)[]
-    | {
-        [attributeName: string]: string;
-      }
-    | string;
-};
-
-type IXMLParserResult = IXMLParserChild[];
+import {
+  parse as parseICU,
+  TYPE,
+  MessageFormatElement,
+  PluralElement,
+  LiteralElement,
+} from '@formatjs/icu-messageformat-parser';
+import { IXMLParserChild, IXMLParserResult } from '../models/fast-xml-parser';
+import {
+  IXliffTag,
+  IXliff,
+  XliffTagName,
+  TAG_NAMES,
+  IXliffInterpolation,
+  IXliffPlural,
+} from '../models/xliff';
 
 export class XliffParser {
-  private tagNames = [
-    'x',
-    'text',
-    'target',
-    'source',
-    'note',
-    'trans-unit',
-    'body',
-    'context',
-    'context-group',
-    'file',
-    'xliff',
-  ];
+  private parser: XMLParser;
 
-  public parse(data: string | undefined): IXliff | undefined {
+  constructor() {
+    this.parser = this.getParser();
+  }
+
+  public parse<T = IXliff>(data: string | undefined): T | undefined {
     if (!data) {
       return undefined;
     }
@@ -47,10 +33,9 @@ export class XliffParser {
       throw validation.err;
     }
 
-    const parser = this.getParser();
-    const parsedXML: IXMLParserResult = parser.parse(data);
+    const parsedXML: IXMLParserResult = this.parser.parse(data);
 
-    return this.convertNodes(parsedXML)[0] as any;
+    return this.convertNodes(parsedXML)[0] as T;
   }
 
   private getParser(): XMLParser {
@@ -60,6 +45,19 @@ export class XliffParser {
       ignoreAttributes: false,
       attributeNamePrefix: '',
       allowBooleanAttributes: true,
+      trimValues: false,
+      stopNodes: [
+        'xliff.file.body.trans-unit.source',
+        'xliff.file.body.trans-unit.target',
+      ],
+      tagValueProcessor(tagName, tagValue) {
+        // Remove spaces only for text node without any other char
+        if (tagValue.trim().length === 0) {
+          return '';
+        }
+
+        return tagValue;
+      },
     });
   }
 
@@ -85,6 +83,16 @@ export class XliffParser {
       return node['#text'] as string;
     }
 
+    if (tagName === 'source' || tagName === 'target') {
+      return {
+        $: node[':@'] ?? {},
+        children: this.convertSourceOrTarget(
+          (node[tagName] as any[])[0]?.['#text']
+        ),
+        name: tagName,
+      };
+    }
+
     if (!this.isValidTagName(tagName)) {
       return undefined;
     }
@@ -96,7 +104,108 @@ export class XliffParser {
     };
   }
 
+  private convertSourceOrTarget(
+    text: string | undefined
+  ): (string | IXliffInterpolation | IXliffPlural)[] {
+    if (!text) {
+      return [];
+    }
+
+    return this.parseICU(text);
+    // if (text.includes('VAR_PLURAL')) {
+    // }
+
+    // return this.parseInterpolation(text);
+  }
+
+  private parseICU(
+    text: string
+  ): (string | IXliffInterpolation | IXliffPlural)[] {
+    const protectInterpolations = this.protectInterpolations(text);
+    const parsedText = parseICU(protectInterpolations.replaced);
+
+    return this.convertICUItems(
+      parsedText,
+      protectInterpolations.interpolations
+    );
+  }
+
+  private convertICUItems(
+    items: MessageFormatElement[],
+    interpolations: string[]
+  ): (string | IXliffInterpolation | IXliffPlural)[] {
+    return items.reduce((acc, item) => {
+      switch (item.type) {
+        case TYPE.plural:
+          return [...acc, this.convertICUPluralItem(item, interpolations)];
+        case TYPE.literal:
+          return [...acc, ...this.convertICULiteralItem(item, interpolations)];
+        default:
+          return acc;
+      }
+    }, [] as (string | IXliffInterpolation | IXliffPlural)[]);
+  }
+
+  private convertICUPluralItem(
+    item: PluralElement,
+    interpolations: string[]
+  ): IXliffPlural {
+    // Get all the options
+    return {
+      counters: Object.fromEntries(
+        Object.entries(item.options).map(([key, value]) => [
+          key,
+          this.convertICUItems(value.value, interpolations),
+        ])
+      ),
+      name: 'plural',
+    };
+  }
+
+  private convertICULiteralItem(
+    item: LiteralElement,
+    interpolations: string[]
+  ): (string | IXliffInterpolation)[] {
+    return this.parseInterpolation(
+      this.restoreInterpolations(item.value, interpolations)
+    );
+  }
+
+  private parseInterpolation(text: string): (string | IXliffInterpolation)[] {
+    const parsedInterpolation = this.parser.parse(`<plural>${text}</plural>`)[0]
+      .plural;
+
+    return this.convertNodes(parsedInterpolation) as any;
+  }
+
+  private protectInterpolations(text: string): {
+    source: string;
+    replaced: string;
+    interpolations: string[];
+  } {
+    const interpolations: string[] = [];
+
+    return {
+      source: text,
+      replaced: text.replace(/<x[^>]+>/g, (value) => {
+        const index = interpolations.length;
+        interpolations.push(value);
+        return `__${index}__`;
+      }),
+      interpolations,
+    };
+  }
+
+  private restoreInterpolations(
+    text: string,
+    interpolations: string[]
+  ): string {
+    return interpolations.reduce((acc, current, index) => {
+      return acc.replace(`__${index}__`, current);
+    }, text);
+  }
+
   private isValidTagName(tagName: string): tagName is XliffTagName {
-    return this.tagNames.includes(tagName.replace('#', ''));
+    return TAG_NAMES.includes(tagName.replace('#', '') as any);
   }
 }