+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### BREAKING CHANGES
+
+- Plural source and target are now parsed into a IXliffPlural
+
+### Changed
+
+- Use fast-xml-parser instead of xml-builder
+
 ## [0.4.0] - 2023-04-27
 
 ### Changed